+++ recipe/teamfolder/member/list.go
@@ -2,6 +2,7 @@ package member
 
 import (
 	"github.com/watermint/toolbox/domain/common/model/mo_filter"
+	"github.com/watermint/toolbox/domain/common/model/mo_string"
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_conn"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_sharedfolder_member"
 	"github.com/watermint/toolbox/domain/dropbox/service/sv_member"
@@ -27,6 +28,7 @@ type List struct {
 	NoMember           rp_model.RowReport
 	Folder             mo_filter.Filter
 	MemberType         mo_filter.Filter
+	ScanTimeout        mo_string.SelectString
 	memberTypeInternal mo_sharedfolder_member.FolderMemberFilter
 	memberTypeExternal mo_sharedfolder_member.FolderMemberFilter
 }
@@ -60,6 +62,10 @@ func (z *List) Preset() {
 		z.memberTypeInternal,
 		z.memberTypeExternal,
 	)
+	z.ScanTimeout.SetOptions(string(uc_teamfolder.ScanTimeoutShort),
+		string(uc_teamfolder.ScanTimeoutShort),
+		string(uc_teamfolder.ScanTimeoutLong),
+	)
 }
 
 func (z *List) Exec(c app_control.Control) error {
@@ -75,7 +81,7 @@ func (z *List) Exec(c app_control.Control) error {
 		return err
 	}
 
-	teamFolderScanner := uc_teamfolder.New(c, z.Peer.Context())
+	teamFolderScanner := uc_teamfolder.New(c, z.Peer.Context(), uc_teamfolder.ScanTimeoutMode(z.ScanTimeout.Value()))
 	teamFolders, err := teamFolderScanner.Scan(z.Folder)
 	if err != nil {
 		return err